refactor: rename utils to misc_models
